requirements.txt
dspy-ai
datasets